web/bean-cafe/challenge/app/requirements.txt
Pillow==9.0.1
Flask-Limiter
transformers